+++ BagBrother/core/api/settings.lua
@@ -79,7 +79,7 @@ local ProfileDefaults = {
 	vault = SetDefaults({
 		borderColor = {1, 0, 0.98, 1},
 		point = 'LEFT',
-		columns = 10,
+		columns = 16,
 		x = 95
 	}, FrameDefaults),
 
@@ -100,11 +100,13 @@ function Settings:OnEnable()
         version = Addon.Version,
 		profiles = {},
 
-		display = {banker = true, guildBanker = true, voidStorageBanker = true, crafting = true, tradePartner = true, socketing = true, auctioneer = true, merchant = true, mailInfo = true, scrappingMachine = true},
 		resetPlayer = true, flashFind = true, tipCount = true, serverSort = true,
+		display = {
+			banker = true, guildBanker = true, voidStorageBanker = true, crafting = true, tradePartner = true, socketing = true,
+			auctioneer = true, merchant = true, mailInfo = true, scrappingMachine = true},
 
 		glowAlpha = 0.5,
-		glowQuality = true, glowNew = true, glowQuest = true, glowSets = true, glowUnusable = true,
+		glowQuality = true, glowNew = true, glowQuest = true, glowSets = true, glowUnusable = true, glowPoor = true,
 
 		slotBackground = 2, colorSlots = true,
 		normalColor = {1, 1, 1},